chore: remove gulp from build process where possible
This makes builds a bit faster because we don't need to re-compile
TypeScript in a few places, and overall sheds some dependencies and cruft.

NOTE: I changed the current `tsc` script in garden-service to `watch`,
and `tsc` in turn only builds once and doesn't watch. I think this is a
bit clearer for developers anyway.
