+++ garden-service/gulpfile.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright (C) 2018 Garden Technologies, Inc. <info@garden.io>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+const { resolve, join } = require("path")
+
+const gulp = require("gulp")
+const pegjs = require("gulp-pegjs")
+
+const pegjsSources = resolve(__dirname, "src", "*.pegjs")
+const destDir = resolve(__dirname, "build")
+
+gulp.task("pegjs", () =>
+  gulp.src(pegjsSources)
+    .pipe(pegjs({ format: "commonjs" }))
+    .pipe(gulp.dest(join(destDir, "src"))),
+)
+
+gulp.task("pegjs-watch", () =>
+  gulp.watch(pegjsSources, gulp.series(["pegjs"])),
+)
