+++ garden-service/package.json
@@ -23,7 +23,6 @@
   ],
   "dependencies": {
     "@kubernetes/client-node": "0.10.1",
-    "@types/koa-static": "^4.0.1",
     "JSONStream": "^1.3.5",
     "analytics-node": "3.3.0",
     "ansi-escapes": "^4.1.0",
@@ -119,7 +118,6 @@
     "@types/dockerode": "^2.5.16",
     "@types/execa": "^0.9.0",
     "@types/fs-extra": "^7.0.0",
-    "@types/gulp": "^4.0.6",
     "@types/has-ansi": "^3.0.0",
     "@types/inquirer": "6.0.1",
     "@types/joi": "^14.3.3",
@@ -131,6 +129,7 @@
     "@types/koa-bodyparser": "^4.2.2",
     "@types/koa-mount": "^3.0.1",
     "@types/koa-router": "7.0.40",
+    "@types/koa-static": "^4.0.1",
     "@types/koa-websocket": "^5.0.3",
     "@types/lodash": "^4.14.126",
     "@types/minimatch": "^3.0.3",
@@ -155,11 +154,7 @@
     "amplitude-js": "^5.2.2",
     "chai": "^4.2.0",
     "gulp": "^4.0.2",
-    "gulp-cached": "^1.1.1",
     "gulp-pegjs": "^0.1.0",
-    "gulp-sourcemaps": "^2.6.5",
-    "gulp-tslint": "^8.1.4",
-    "gulp-typescript": "^5.0.1",
     "handlebars": "^4.1.2",
     "is-subset": "^0.1.1",
     "linewrap": "^0.2.1",
@@ -185,20 +180,23 @@
     "typescript-formatter": "^7.2.2"
   },
   "scripts": {
-    "build": "gulp build",
-    "build-ci": "gulp build",
+    "add-version-files": "node ./build/src/bin/add-version-files.js",
+    "build": "npm run tsc && gulp pegjs && npm run add-version-files && npm run generate-docs",
+    "build-ci": "npm run build",
     "build-container": "npm run build && docker build -t gardendev/garden:latest .",
     "check-package-lock": "git diff-index --quiet HEAD -- package-lock.json || (echo 'package-lock.json is dirty!' && exit 1)",
     "clean": "shx rm -rf build",
-    "dev": "npm link && npm run tsc",
+    "dev": "npm link && npm run watch",
     "dist": "npm run build",
+    "generate-docs": "node ./build/src/bin/generate-docs.js",
     "integ": "./test/integ/run",
     "integ-full": "node ./build/test/integ/integ-full.js",
     "prepare": "npm run snyk-protect",
     "snyk-protect": "snyk protect",
     "test": "node_modules/.bin/mocha",
-    "tsc": "tsc -w -p . --outDir build --declaration --declarationMap --sourceMap",
-    "view-report": "open coverage/index.html"
+    "tsc": "tsc -p . --outDir build --declaration --declarationMap --sourceMap",
+    "view-report": "open coverage/index.html",
+    "watch": "npm run tsc -- -w"
   },
   "pkg": {
     "scripts": [
